Add IAppUi.
