+++ src/ui/CMakeLists.txt
@@ -3,13 +3,14 @@ set(NAME "Ui")
 include(${PROJECT_SOURCE_DIR}/cmake/StaticAnalyzers.cmake)
 
 add_library(${NAME} STATIC
-  "Ui/AppUi.h"
+  "Ui/IAppUi.h"
   "Ui/ImageInfo.h"
-
-  "Ui/AppUi.cpp"
+  "Ui/src/AppUi.h"
+  "Ui/src/AppUi.cpp"
 )
 
 target_include_directories(${NAME} PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_include_directories(${NAME} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/Ui/src)
 target_include_directories(${NAME} PRIVATE ${PROJECT_SOURCE_DIR}/src/core)
 
 target_compile_features(${NAME} PRIVATE cxx_std_20)
